feat. quantity in detailorder changed to decimal

+++ src/shared/infra/typeorm/migrations/1597895221852-CreateOrdersDetail.ts
@@ -41,7 +41,7 @@ export default class CreateOrdersDetail1597895221852
           },
           {
             name: 'quantity',
-            type: 'int'
+            type: 'decimal'
           },
           {
             name: 'created_at',
